GenAI Takeaway Assignment/requirements.txt
fastapi
uvicorn
pandas
faiss-cpu
rank-bm25
scikit-learn
python-dotenv
beautifulsoup4
pydantic
sentence-transformers
torch
cohere
